Moved collaborator components into publisher app
