+++ static/js/src/publisher-admin/hooks/index.ts
@@ -1,3 +1,13 @@
+import useCollaboratorsQuery from "./useCollaboratorsQuery";
+import useInvitesQuery from "./useInvitesQuery";
 import usePackage from "./usePackage";
+import useRevokeMutation from "./useRevokeMutation";
+import useSendMutation from "./useSendMutation";
 
-export { usePackage };
+export {
+  useCollaboratorsQuery,
+  useInvitesQuery,
+  usePackage,
+  useRevokeMutation,
+  useSendMutation,
+};
