+++ static/js/src/publisher-admin/hooks/useCollaboratorsQuery.ts
@@ -1,0 +1,23 @@
+import { useQuery } from "react-query";
+
+function useCollaboratorsQuery(packageName: string | undefined) {
+  return useQuery("collaboratorsData", async () => {
+    const response = await fetch(`/api/packages/${packageName}/collaborators`, {
+      cache: "no-cache",
+    });
+
+    if (!response.ok) {
+      throw new Error("There was a problem fetching collaborators");
+    }
+
+    const collaboratorsData = await response.json();
+
+    if (!collaboratorsData.success) {
+      throw new Error(collaboratorsData.message);
+    }
+
+    return collaboratorsData.data;
+  });
+}
+
+export default useCollaboratorsQuery;
