+++ static/js/src/publisher-admin/pages/Collaboration/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Collaboration";
