+++ static/js/src/publisher-admin/types/components/Collaborator.d.ts
@@ -1,0 +1,9 @@
+type Collaborator = {
+  account: Publisher;
+  "created-at": string;
+  "created-by": Publisher;
+  permissions: Array<string>;
+  "updated-at": string | null;
+};
+
+export default Collaborator;
