+++ static/js/src/publisher-admin/types/components/Invite.d.ts
@@ -1,0 +1,12 @@
+type Invite = {
+  "accepted-at": string | null;
+  "accepted-by"?: Publisher;
+  "created-at": string;
+  "created-by": Publisher;
+  email: string | null;
+  "expires-at": string | null;
+  "invite-type": string | null;
+  "revoked-at": string | null;
+};
+
+export default Invite;
