+++ static/js/src/publisher-admin/types/components/Publisher.d.ts
@@ -2,7 +2,8 @@ type Publisher = {
   "display-name": string;
   id: string;
   username: string;
-  validation: string;
+  validation?: string;
+  email?: string;
 };
 
 export default Publisher;
