+++ static/js/src/publisher-admin/types/global.d.ts
