+++ static/js/src/publisher-admin/types/index.d.ts
@@ -1,4 +1,6 @@
+import Collborator from "./components/Collaborator";
+import Invite from "./components/Invite";
 import Package from "./components/Package";
 import Publisher from "./components/Publisher";
 
-export { Package, Publisher };
+export { Collaborator, Invite, Package, Publisher };
