+++ static/js/src/publisher-admin/utils/__tests__/getFilteredCollaborators.test.ts
@@ -1,4 +1,4 @@
-import getFilteredCollaborators from "./getFilteredCollaborators";
+import getFilteredCollaborators from "../getFilteredCollaborators";
 
 const collaborators = [
   {
