+++ static/js/src/publisher-admin/utils/__tests__/getFilteredInvites.test.ts
@@ -1,4 +1,4 @@
-import getFilteredInvites from "./getFilteredInvites";
+import getFilteredInvites from "../getFilteredInvites";
 
 const invites = [
   {
