+++ static/js/src/publisher-admin/utils/__tests__/getInvitesByStatus.test.ts
@@ -1,6 +1,6 @@
 import { add } from "date-fns";
 
-import getInvitesByStatus from "./getInvitesByStatus";
+import getInvitesByStatus from "../getInvitesByStatus";
 
 const invites = [
   {
