+++ static/js/src/publisher-admin/utils/__tests__/publisherMatchesFilterQuery.test.ts
@@ -1,4 +1,4 @@
-import publisherMatchesFilterQuery from "./publisherMatchesFilterQuery";
+import publisherMatchesFilterQuery from "../publisherMatchesFilterQuery";
 
 const publisher = {
   "display-name": "John Doe",
