+++ static/js/src/publisher-admin/utils/getInvitesByStatus.ts
@@ -1,14 +1,6 @@
-import { isAfter } from "date-fns";
+import { isPending, isExpired, isRevoked } from "../utils";
 
 import type { Invite } from "../types";
-
-export const isAccepted = (invite: Invite) => invite?.["accepted-at"] !== null;
-export const isRevoked = (invite: Invite) => invite?.["revoked-at"] !== null;
-export const isExpired = (invite: Invite) =>
-  invite?.["expires-at"] !== null &&
-  isAfter(new Date(), new Date(invite?.["expires-at"]));
-export const isPending = (invite: Invite) =>
-  !isAccepted(invite) && !isRevoked(invite) && !isExpired(invite);
 
 function getInvitesByStatus(
   invites: Array<Invite>,
