+++ static/js/src/publisher-admin/utils/index.ts
@@ -1,4 +1,29 @@
 import capitalize from "./capitalize";
 import languages from "./languages";
+import getFilteredCollaborators from "./getFilteredCollaborators";
+import getFilteredInvites from "./getFilteredInvites";
+import getInvitesByStatus from "./getInvitesByStatus";
+import publisherMatchesFilterQuery from "./publisherMatchesFilterQuery";
+import buildInviteTableRows from "./buildInviteTableRows";
+import isAccepted from "./isAccepted";
+import isExpired from "./isExpired";
+import isPending from "./isPending";
+import isRevoked from "./isRevoked";
+import RecoilObserver from "./RecoilObserver";
+import QueryProvider from "./QueryProvider";
 
-export { capitalize, languages };
+export {
+  capitalize,
+  languages,
+  getFilteredCollaborators,
+  getFilteredInvites,
+  getInvitesByStatus,
+  publisherMatchesFilterQuery,
+  buildInviteTableRows,
+  isAccepted,
+  isExpired,
+  isPending,
+  isRevoked,
+  RecoilObserver,
+  QueryProvider,
+};
