+++ static/js/src/publisher-admin/utils/isAccepted.ts
@@ -1,0 +1,5 @@
+import type { Invite } from "../types";
+
+const isAccepted = (invite: Invite) => invite?.["accepted-at"] !== null;
+
+export default isAccepted;
