+++ static/js/src/publisher-admin/utils/isExpired.ts
@@ -1,0 +1,9 @@
+import { isAfter } from "date-fns";
+
+import type { Invite } from "../types";
+
+const isExpired = (invite: Invite) =>
+  invite?.["expires-at"] !== null &&
+  isAfter(new Date(), new Date(invite?.["expires-at"]));
+
+export default isExpired;
