+++ static/js/src/publisher-admin/utils/isRevoked.ts
@@ -1,0 +1,5 @@
+import type { Invite } from "../types";
+
+const isRevoked = (invite: Invite) => invite?.["revoked-at"] !== null;
+
+export default isRevoked;
