--- static/js/src/publisher/collaboration/index.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { createRoot } from "react-dom/client";
-
-import App from "./components/App";
-
-const container = document.getElementById("root");
-const root = createRoot(container!);
-root.render(<App />);
